Refactored xtea into test and header files
