+++ resources/c/Makefile
@@ -1,6 +1,6 @@
 all:
 	gcc stego_engine.c test_stego.c -o stego -Wall
-	gcc xtea_engine.c -o xtea -Wall
+	gcc xtea_engine.c test_xtea.c -o xtea -Wall
 clean:
 	rm stego
 	rm xtea
